+++ app/Http/Controllers/Backsite/RoleController.php
@@ -42,6 +42,8 @@ class RoleController extends Controller
 	 */
 	public function index()
 	{
+		abort_if(Gate::denies('role_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 		$role = Role::orderBy('created_at', 'desc')->get();
 
 		return view('pages.backsite.management-access.role.index', compact('role'));
@@ -75,6 +77,8 @@ class RoleController extends Controller
 	 */
 	public function show(Role $role)
 	{
+		abort_if(Gate::denies('role_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 		// contains all the Permissions associated with the Role
 		$role->load('permission');
 
@@ -86,6 +90,8 @@ class RoleController extends Controller
 	 */
 	public function edit(Role $role)
 	{
+		abort_if(Gate::denies('role_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 		$permission = Permission::all();
 
 		// contains all the Permissions associated with the Role
@@ -114,6 +120,7 @@ class RoleController extends Controller
 	 */
 	public function destroy(Role $role)
 	{
+		abort_if(Gate::denies('role_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
 		$role->forceDelete();
