+++ app/Http/Controllers/Backsite/SpecialistController.php
@@ -40,6 +40,8 @@ class SpecialistController extends Controller
 	 */
 	public function index()
 	{
+		abort_if(Gate::denies('specialist_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 		$specialist = Specialist::orderBy('created_at', 'desc')->get();
 
 		return view('pages.backsite.master-data.specialist.index', compact('specialist'));
@@ -73,6 +75,8 @@ class SpecialistController extends Controller
 	 */
 	public function show(Specialist $specialist)
 	{
+		abort_if(Gate::denies('specialist_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 		return view('pages.backsite.master-data.specialist.show', compact('specialist'));
 	}
 
@@ -81,6 +85,7 @@ class SpecialistController extends Controller
 	 */
 	public function edit(Specialist $specialist)
 	{
+		abort_if(Gate::denies('specialist_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
 		return view('pages.backsite.master-data.specialist.edit', compact('specialist'));
 	}
@@ -105,6 +110,8 @@ class SpecialistController extends Controller
 	 */
 	public function destroy(Specialist $specialist)
 	{
+		abort_if(Gate::denies('specialist_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
 		$specialist->forceDelete();
 
 		alert()->success('Success Message', 'Successfully deleted specialist');
